requirements.txt
aiofiles==24.1.0
aiogram==2.25.1
aiohappyeyeballs==2.6.1
aiohttp==3.8.6
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
frozenlist==1.6.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
magic-filter==1.0.12
multidict==6.4.4
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
pytz==2025.2
rfc3986==1.5.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
yarl==1.20.0
